Storing grid height, width and step size and using them to position new balls.
Using the position of previous ball to position new balls can be an issue if the user taps on the screen while the old
ball is being moved. Also storing the values speeds up the positioning because the grid array doesn't need to be looked
up to get the new positions.